requirements.txt
Django==5.0.10
django-storages==1.14.2
azure-storage-blob==12.19.0
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0